requirements.txt
langchain_community
python-dotenv
langchain_openai
langchain_core
langchain_google_vertexai
langchain
streamlit
langchain-google-genai
langchain-anthropic
langchain-groq
transformers